New unique: Parghit the Mighty
Following the example of Maggie/Margery, Parghit is an "alt-timeline"
version of Pargi. Parghit appears in late Depths, Zot, Crypt and Tomb,
and sports high-enchantment golden dragon scales (with a 1% chance of
wearing the Dragon King) and an astonishing 27 HP regenerated per turn.
He's definitely killable, but best not underestimated.

I considered also giving him large rocks. It seemed rude, but I might
go back to that if it seems appropriate.

Credit Hellmonk for several parts of this idea.