New unique: Josephina the Ice-Lich
As one more 'alt-timeline' unique, Josephina is an upgraded version of
Josephine. She appears in late Vaults, Depths, Crypt, Tomb and Zot, and
borrows Wendigo Seracfall (turning simulacrula into 3d33 bombs) to
add to her classic Ghostly Fireball. She comes with a guaranteed randart
ring of ice.

I suspect a lot of the numbers here will need more fine-tuning. She
might be a bit of a glass cannon at the moment, especially with the rF-,
and I think the placement range is probably a bit off.