fix: <enter> on the spellcasting menu casts the last spell again
This was broken for both menu_arrow_control = true and false. The
previous behavior was restored if arrow control is off; if arrow
control is on, the menu automatically defaults to hovering over
the last-cast spell, which means that pressing enter will cast that
spell.

Either way, the last-cast spell is shown with a '+' after the hotkey
instead of a '-'. This does not restore the behavior when the player
knows only one spell; that was mostly for newbies who are presumably
keeping the default of true for arrow control.

(cherry picked from commit 9887b6b0900eb370c2c57c4beeab5cc65fea7280)