fix: select all menu items when pressing a type shortcut
Because the second iteration of the loop did not have the
MF_SINGLESELECT check, using a non-primary hotkey to select things you
had already scrolled past did not reliably select all items.

The best way to reproduce this was to have several weapons in your
inventory, scroll down past them on the (d)rop screen, and then press ).
The topmost weapon would not be selected. (More specifically, I suspect
it was being selected twice, leading to it being deselected.)