feat: Update a Depths float vault
For `gammafunk_depths_water_palace`, update the monster list to be more
current with the Depths roster. Instead of solo water elementals at
certain locations, use merfolk exclusively, but have a chance for two of
these locations near the central chamber entrance to be elemental
wellsprings. Reglyph a bit as necessary.