fix: don't double prompt for monsters with Yara's (12669)
Using stop_attack_prompt and firing a tracer is overkill. The former
uses a better verb and works well with modern hitfuncs. The latter
includes a self-damage check. It would be great if the former could be
converted but that's not a small task and is likely to introduce a bunch
of new double prompt bugs. For now I've gone with checking for
self-damage separately.

(cherry picked from commit 0a91294c2cb7e035c3a0da177705b9894bfcaeb7)