fix: scroll more sanely in webtiles
Page up/down was overshooting. This commit seems to bring it to a better
place (although the interaction with the shade is messy). I'm not sure
why this code was previously commented, as far as I can tell (aside from
compensating for the shade) it was correct, so hopefully there isn't
some problem I didn't find.

Resolves https://crawl.develz.org/mantis/view.php?id=12193