Reduce Wucad Mu's channeling effectiveness (atomjack)
The new formulae resulted in Wucad Mu scaling much more strongly than
intended, up to approximately a 99% chance to make spellcasting free at 27
Evocations skill. Ideally the formulae will be reworked in the longer term
to also make the effect scale better for regular orbs of energy at lower
skill, but in the meantime reduce the ceiling by counting Wucad Mu as 1.5
sources of channeling instead of 2, and removing the stacking effect of
multiple sources on the backfire rate.