fix: handle webtiles SIGHUP messaging differently
SIGHUP indicates that something killed the process. While this can
happen for various reasons, it mostly happens on some form of
disconnect. If the player is around to see this message (as in Mantis
11595), they probably hit the back button or something, and it's
misleading to describe this as a crash ("Unfortunately your game
terminated due to an error."). In any case, there's no crash report on a
SIGHUP. This commit adds a new message in client.js that'll require
caches to catch up, but until then the default case message is fine, and
prints the exit reason.

More or less resolves https://crawl.develz.org/mantis/view.php?id=11595

(cherry picked from commit 1d65ae1bb44c678020c7d57118051b73341f63fd)