fix: adjust mouse hover suppress timing (ebering)
This was mainly tested in firefox, which takes about 10-20ms to deliver
mouseevents to the menu after a scroll, so 80ms seemed nice and
conservative. However, in Safari, the timing seems to be more on the
order of 100ms. (Maybe it's intentionally delaying them if they result
from rendering rather than actual mouse movement?) I've upped this
timeout to 200ms. Even with a value that might in principle be
noticeable, I wasn't able to generate any intentional physical motion
that revealed the suppression when trying to use a mouse.

Unfortunately this change does make it a lot easier for the timeouts to
overlap, which a simple bool doesn't handle very well, leading to the
more involved approach here.

(cherry picked from commit e7179d36ccb5ef40561218a7cc49d68a3fbac9c4)