fix: update some mac plist stuff
I have no idea why LSRequiresNativeExecution was previously being
ignored (the 0.27.1 release has this set and runs fine under Rosetta),
but as of the 0.28-b1 release artifacts, it is *not* being ignored and
this prevents it from running under Rosetta. I'm not sure why this was
even set in the first place.

I've also adjusted the minimum version field to what we actually build
for. Probably more values here need a pass, e.g. the bundle id is pretty
outdated.