fix: default runrest ignore transformation ending messages
As reported in #2189 the default options repeatedly interrupt resting
for transmuters, and this is particularly noticable in the early game
when spell power is low. This commit adds a default ignore for all
transformation-expiring-soon messages and for transformation end
messages for all good forms.

Wisp, Bat, Pig, and Tree form still interrupt rest, as these are far
more likely for the player to want to rest off.

This commit does entail a tradeoff: it makes the experience smoother for
newer players in the early game at the expense of requiring a player who
wants to remain in a form while auto-exploring/auto-traveling to edit
their rc.

It's not clear to me that there's a way to easily satisfy both use
cases, someone will need to edit their rc. This change shifts the onus
to do so to what are probably more experienced players.

Closes #2189