Give kobolds +3 xbows
As something they're better at than other races (being mechanically
inclined), it's nice to accentuate that strength a little more than
their former +2. Plus, it's funny to give a race that has a big
obvious disadvantage for ranged combat (reduced LOS) a great apt for
it - creates a different kind of experience!