fix: invert some mistaken transparency logic
This seems to be the opposite of what was intended in 7e0d850643. On top
of the wrong transparency logic, the fact that these vaults can get a
transparent tag added if they are used in vaults was generating false
positives in seed testing (where the transparent tag persists to a later
run). I'm not actually sure its great for the vaults layout code to
force-add `transparent` like that (even though I probably added it), but
I'm leaving that alone for now.

(cherry picked from commit c37658a0fea4e6c0196d6b1b5df7965af041e6c1)