fix: regularize confused self-attacks (perryprog)
Previously, there was basically one way where a monster could
accidentally hit itself while confused (with a number of other rare
special cases): if it targeted another monster, `mmov` would be reset
after the `fight_melee` call, and then the later there's an additional
1/6 chance for a self-attack. There's code that appears intended to
handle a similar player case in `_monster_move`, as well as a chance on
top of the monster rolling mmov as 0,0, but these were unreachable due to
variously an earlier batch of code that handles attacking the player,
and an early return if `mmov.origin()` is true. This commit regularizes
things to what appears to be intended. Further tweaks may be justified,
and in particular I'm not sure the purely random chance of
self-attacking is needed. After this commit:

2/3 chance of stumbling, with a 1/n chance of rolling 0,0 (n=habitable
spaces), and then a 1/6 chance of a self-attack. So *probably* rare
enough to be not worth kiting to get.

If the monster targets another monster, or (now fixed) the player: a 1/6
chance of a self-attack on top of the swing. Oddly (?), this doesn't
change energy and the monster always gets a swing for this case, so they
attack twice for the usual energy cost. As noted in the comments, I'm
interpreting this as a kind of slapstick effect. But it might be more
consistent if that 1/6 roll attacked themselves *instead* of the
intended target, not sure.