Change the HUD "heavily drained" message from "drained" to "drained!".
As things were, both light and heavy drain were displayed as "drained", meaning that you could be near "3 orcs (1 drained, 2 drained)". Changing one to "drained!" tells the player which is which.

The only other message which is used for more than one status is "clouds", and a type of monster can only be surrounded by one type of cloud.