Disable the MCC targeter by default
It seemed to mostly confuse players by leading them to think that the
spell was targeted, since they could move the cursor around while in
the targeter. It's also not super useful, since whatever monster is
highlighted when you start the spell may not be the eventual victim.

The MCC targeter can still be seen by quivering and (f)iring the spell,
or by adjusting the rcfile.