fix: stabilize serpent of hell placement
The previous code was a bit of a pain to understand, and also seems to
have never been updated for pregeneration. This simplifies the
calculation, and restructures it to not rely on player exploration. It
is still entirely possible for hell end vaults to mess this up.

(cherry picked from commit 25f29fbe2000b95f283948288b95a75ba9b0fc5d)