fix: check ban flag for token logins
I think I didn't add this before because it is awkward to check at the
point of login without adding another db call. However, user flags are
collected from the db shortly after the point of checking the token, so
this is a perfectly workable point to check the ban flag again. (Some of
this code could perhaps be consolidated.)