feat: various account management enhancements
Nick restriction improvements:
* Allow defining a custom name check function in config. Some things are
  a bit easier to do this way than with a regex.

Two restricted registration modes:
* set `new_accounts_disabled` to disable new account creation
  altogether. This is a bit quick and dirty, and users currently don't
  get an error until they hit the submit button.
* set `new_accounts_hold` to allow new account creation, but mark any
  new accounts with an "account hold" flag. This flag:
    - prevents login via ssh, where relevant. (It just sets the ban
      flag, so messaging to players may be confusing..)
    - prevents the account's games from appearing in the lobby or being
      spectated.
    - prevents the account from using chat.
    - prevents the account from spectating altogether. (Of course they
      can log out and spectate as anonymous.)
    - Players with this flag can still play games.
    - This flag is implemented as a new dgamelaunch db flag. Since dgl
      itself hasn't added any flags in like 15 years, hopefully this is
      safe.

The account hold flags can be set and cleared via the wtutil.py command
line tool. I'd like to add an admin panel interface too, but that isn't
done. Ths flag, if set, has an effect even when the new account hold
mode isn't enabled, so it can also be used as a kind of soft ban if
someone is causing trouble in chat but for whatever reason isn't judged
to merit an outright ban.