feat: smooth early super ood cap and document
In des syntax the glyph `8` means depth * 2 + 4. A "super ood". The
glyph `9` means depth + 5, a "moderate ood". Both of these are
constrained in D:1-8 at present, the maximum vault OOD is depth * 2 over
D:1-8. This cap expires after D:5 for `9`, but continues to impact the
meaning of `8` until D:8.

The result is that from D:8 to D:9 there's a large jump in difficulty of
vaults. This commit smooths things, so that the additive term goes up
by 1 for each floor afte D:8 until reaching its full value by D:12 (at
which point `8` means max depth since the cap of D's OOD table is 27).