fix: allow disabling yaml games definitions
Since these aren't templated on dgl setups, they are a bit of a pain to
deal with even if the goal is just to remove the default one from trunk
or prevent its installation. This commit provides another approach,
where you can set the directory to `None` and skip loading yaml game
definitions altogether.