fix: check webtiles db flags more consistently and act on them
This improves the handling of using the cli to change the ban/hold state
of players who are currently logged in. The main change to db access is
that the flags are updated when starting a game, and changes (in either
direction) should take effect immediately. To get this to work I had to
improve some of the logout handling (which previously relied on a
reload, and couldn't be triggered from the server). Hopefully the I/O
cost of this extra db check on I/O-bound servers isn't impactful, we'll
see.

This commit also fixes a bug where a held account that tries to spectate
got the error info through the wrong message (well, the right one didn't
exist yet) causing the user/password boxes to reappear.