feat: add a CLO to prevent writing player bones
This is not really intended as a player facing feature (it doesn't
prevent ghosts from being generated or anything like that), but as an
option to support community management features on webtiles.
Implementing as a separate commit to make it easier to backport.

(cherry picked from commit abffb2d2b2ae4bdebd4207ad2a5038eda789bcbf)