feat: a means of preventing player ghosts from held accounts
This is implemented a bit generally for maximum compatibility, but
because it uses the new `-no-player-bones` option, a held account will
not be able to start games with crawl binaries that do not support this.
This applies both to old versions and forks. I will cherry-pick the
options commit to a few stable versions back. Should have no impact on
servers that aren't using account holds.