Fix Shilleiagh noise (frdfg)
When Devastator was added in 9cc3ee3f91d, it was intended to make 7
noise with every hit. However, since it was an explosion, it accidentally
made 15 noise instead, and hence the wiki noted it as making 'fireball-level
noise'. This continued for 10 years until 58ad9d9fa3b, which accidentally
muted it while trying to make Irradiate silent.

This commit restores Devastator to its intended 7 noise per hit.