Remove identification scrolls from troves
In the past year, identification scrolls have become increasingly less
important as gear now is automatically identified, and never cursed.

By the time players get into a trove, they generally will have no
further use for scrolls of identify, so those should be removed from the
good_scroll table. In their place, we now have immolation scrolls, which
are a fun tool for safely fighting large armies, and torment scrolls,
so that these troves can be just as fun for demons/undead as they are
for other species. These two additions are equally as common as scrolls
of holy word.