docs: attempt to despoiler various aux/retaliation muts
I say "attempt" because this stuff is complicated enough that there may
be all sorts of mistakes in here still, or even some random rolls I've
missed entirely. In many cases I have chosen to give slightly round
numbers (e.g. 40% rather than 38%), maybe these should be made exact.
Also, this is a lot of numbers, but I think we have become pro-numbers
enough for this sort of change. This commit began in response to
https://www.reddit.com/r/dcss/comments/svj867/demonic_touch_utterly_useless/
where it became apparent that how this particular mut worked was
completely opaque to players (especially since its behavior is pretty
atypical), but once I started doing this it became clear to me that
basically all of these would be better with more information.

I've tried to get all of the direct combat effects in this change.  Some
of the Ru mutations that are responsive (e.g. no potion heal) could
maybe use similar work.