feat: indicate stateful ?tele behavior in ItemUseMenu
The stateful behavior of scrolls of teleportation is kind of easy to
forget about, and also unique (I think) among consumables. Because of
this, it seems helpful to indicate if it will cancel an existing
teleport explicitly. I've done this in the `r` menu, another place this
could be done is in the item description itself.