fix: webtiles save info collection
This broke around b939be14790 but I hadn't sat down to figure out what
happened until now. (Amazed that there was never an actual issue about
this that I'm aware of!)

It may have been sort of a blessing in disguise as at least on CAO this
callback chain has become pretty heavy and often times out. We'll see...