refactor: clean up `chr` chunk loading
This was sort of impressively convoluted and bad: the tag code read chr
chunk data directly into `you`, and then a `player_save_info` object was
constructed from that. In most cases, this was wrapped in code that made
a copy of `you` and then restored it after getting the info object.
There were also all sorts of inconsistencies in what data was stored in
this object. I'm guessing no one really wanted to mess with this old
code where the cost of breaking things is high, but I'm going to give it
a short here. This commit:

* Cleans up `player_save_info` to exactly match the data in the `chr`
  chunk, aside from version info (and doll info used for tiles main
  menus). Add a few fields, remove one (`experience`) that seems to be
  completely unused.
* Add a constructor to `player_save_info` to handle default values a bit
  more cleanly
* Rewrite the tag code to read directly to a player_save_info object,
  and only copy data into `you` on demand
* more comments, code reordering to match chunk data order, slightly
  better error messages in a few cases