fix: (re)allow launching from `F`
Previously this was implemented via ammo, which is gone. But it seems
very confusing to not have launchers selectable from this menu, which
after this change became limited only to throwing. This commit is a bit
of a quick hack to do this, may need some more work. This means you
can't use `F` any more to toss away a wielded launcher, though this is
still possible via quivering it (with `Q**`). You can still toss a
non-wielded launcher from this menu.

More generally, it might be good to also allow quivering newlaunchers so
that they still work with `f`/`p` (especially the latter, it's confusing
that CMD_AUTOFIRE no longer has a way to fire for this case), but this
will not be as easy so I'm not going to get to it immediately.  The
division of labor for autofight/autofire introduced with quivers may
need some rethinking.