feat: rework simulacrum
Simulacrum is now a smite-targeted spell that places an irresistable status.
A monster that dies with the status will spawn a power-dependent number of
simulacra (no longer depends on monster size). I think this is enough of a
difference to distinguish it from dchan and infestation.

Desperately needs a status icon and better lore.