feat: add an explicit toss submenu
Tossing is currently in a weird state where as more and more things
become quiverable, they have two types of actions that could apply,
making them harder to figure out how to toss. This is all sort of still
in a mess for other menus, but this (quite basic) submenu is guaranteed
to always use `ammo_to_action` and therefore only throw.