feat: readd quivering/firing ranged weapons
It's proven somewhat confusing to many players that the relevant UI
elements no longer worked with launchers after ammo removal. This commit
restores the ability to quiver launchers, allowing them (if quivered) to
be triggered on default keybinds using `f` and `p`, as well as `v` and
tab regardless of what is quivered.

I've restored some corresponding options as well. It's entirely possible
their defaults should be different than they used to be, if these are
better viewed as compatibility options now. But for now, launchers are
in the fire order (this never got removed), and `launcher_autoquiver`
defaults to true.