Update job recs for og/tr (Lici)
Now that you can't start with large rocks as Hu or AM, don't
recommend them to trolls or ogres.

AE background rec for ogre is semi-arbitrary. Seemed to have a
high recent winrate for not excessively dubious reasons (wasn't a
nem choice, wasn't one player streaking, etc). Seems reasonable for
apts and so on (bouncing less dangerous when you have high HP).

Might add Gl as a rec if a-good-start lands.