Mummy reform: Decay ability & necro enhancers
Mummies have long been notoriously awful, and recent versions have
done nothing to help them. That's fine to some extent - it's funny
to have a challenge species - but there's a limit, and it's with
planned changes to add many potions to different starts, mummies are
more left behind than ever.

So, to give em more distinctiveness and early game tactical options,
add a 'decay' ability that lets them plop miasma on their current
tile. This plays a bit like super conjure flame (a wall early, a
damage enhancer later), and is very funny in certain cases (vs
frenzied enemies, Robin, etc).

To limit the effect, Decay costs HP drain, currently about 75% of an
Apocalypse - it's not something you can use every fight. If it's still
too strong, we can raise the costs and/or adjust it to do stat drain
instead, which would line up with the effect of malmut on mummies.

Also move necro enhancers from XL 13 and 26 to 3 and 13, to make them
show up in a reasonably relevant timeframe.