Channeling: improve the worst cases, reduce the best cases
If a player equips the orb of energy with low evocations skill, they'll get constant backfires. Printing a message could help hint that it's not just bad luck, it's a bad idea.

When backfires do happen, due to low skill or low luck, refund the mana anyways - it's being gained at a rather steep and unsustainable price, after all.

Lastly, to reduce the chance of runaway situations with free mana, only refund slightly over half of the spell's mana cost.