fix: handle undefined item slots with action_panel_glyphs=true
This can specifically cause a crash when dropping an artefact and then
picking up something in the same slot, where it tries to get a glyph on
an undefined item_def and artefact-related properties are in an
inconsistent state. Possibly a pseudo-bug that flags are not reset for
this case, or that get_item_glyph tries to do something for this case?
(I think otherwise, this call was mostly was stable.)

Resolves #2434