Ghouls are very one-dimensional due to extremely poor int. Further, their gimmick is more irritating than fun or interesting, and is not very thematic. This branch seeks to fix this.
The current implementation drains the Ghoul's int while an enemy is in LOS;
the rate of drain decreases the more drained the Ghoul is.
Specifically, every 3 to 5 turns, the ghoul loses 1 int with probability
(base int - drain)/base int
Kills at melee range restore 1 int, so long as the Ghoul is the source.

The ghoul's old gimmick is preserved, save that it heals twice as often,
but only at melee range.


Accurate messages are not implemented, and will not be until the
gimmick is more set in stone.
At a bare minimum number changes will be needed.