Add an arrival vault with teleporters
Teleporters generally only show up in dangerous scenarios; this puts
a pair of teleporters into an arrival vault for people to play with.
These can also be used to juke gnolls and panlords, in times of need.

(Inspired by the "throwing tutorial" key holder vault in Brogue, which
shows up early and often in dungeon generation, helpiing the player
learn that items can be thrown to trigger traps from a distance.)