Shift gyre and gimbal to be early-game
The protection brand is particularly relevant when enemies may only be
swinging for 9 to 20 damage. Damage-dealing, too, can be easier for a
short blade when AC is lower. In the late-game, most short blades that
maintain relevance will be sporting different brands. If one finds this
weapon in Vaults, they're probably already wielding something else and
unlikely to switch to Gyre.

Having this weapon crop up earlier will give players a chance to use it
during a time where it's not outclassed by as many other weapons. Making
the base item a short sword may help nudge players to look for
replacements. Also, there aren't any artifact short swords currently.