feat: new monser "Starflower", an abyssal oklob
A stationary plant/tentacle hybrid monster with harpoon shot, reach 3
distortion melee, reach 2 melee, and an additional melee attack. Meant
as a vault-placed monster for abyssal vaults, guarding an exit or a rune
or what have you, as a threat that's dangerous to be close to, but can
be neutralized if the harpoon shot can be blocked or avoided somehow.