buff damage, change attack order
the idea with constriction being the first attack was that it can hold
the player in place for the next two attacks. putting it last instead
helps lean into crawl's [in]famously spiky, swingy damage. if the
monster rolls Blink on the distortion effect table, you may get batted
away quickly for minor damage. but if the monster rolls bonus damage or
nothing, you're in for 3 decently hard-hitting attacks in a row.

(not playtested yet, still WIP)