Simplify job starting stats
Frequently, experienced players will remind newer players that jobs /
backgrounds are meant to be a starting kit, not a lifelong class choice
commitment. One proviso to that is the starting stats. Most roles'
str/int/dex numbers are in the 2-6 range, but there are a few big outliers.

A TrBe begins with 3 intelligence, putting them easily within range for
death by brainlessness. A new player who chooses this combo may not
realize that they'd counterintuitively be wise to increase their int.
And even when the intensely skewed stats of Be or Cj don't put one in
danger of stat-zero, they still result in the background having an
impact that will last well beyond a "starting kit".

This commit greatly reduces the stat-based differences between roles,
which will make it easier for players to "pivot" away from the starting
kit in a way that I believe is more in tune with dcss's style. The risk
of stat-zero will also be greatly reduced. If some roles lose too much
early-game power from this shift, 1-2 points of extra stats could help.