Adjust bend space.
The higher the power, the higher the probability of moving the surrounding objects together, and at the same time, the higher the damage. However, this effect also applies to neutral objects, so players can become hostile to them if they use this power to escape from the existence of abyss that emerged through 'corrupt'. The higher the power, the higher the probability of Blink, and the higher the likelihood of them antagonizing the player. This bizarre interaction should have been eliminated sooner. Now, even if a neutral monsters is damaged through bend space, it is no longer hostile to the player.