vaults: arrival vaults by SpicyCebolla
- spicycebolla_arrival_risky_transporter

This teaches that using a transporter can be risky:
the 'porter takes you from pretty safe terrain to a wide open space.

[ Committer's note: simplified the substitutions, cleaned up, added
  sightlines. ]

- spicycebolla_arrival_clurvy

a few simple 20% chances of theme variations, giving different
iterations on a forgotten stand of trees (possibly long forgotten)

[ Committer's note: adjusted tree colouration, prevented a small chance
  of a teleport closet ]

- spicycebolla_green_bress_n_ham
- spicycebolla_green_bress_n_ham_2

Simple arrivals.

- spicycebolla_arrival_ecumenical

Places a faded altar with stairs (or a very lucky teleport) required to
reach it.

[ Committer's note: This slightly increases the 87.5% chance of a D:1-3
  faded altar, but not so much that it matters. Closes #2492 ]