docs: fix/clarify runrest prepend vs append
The docs are overall not very consistent, but for options where it
matters, there seems to be a preference to teach ^= as the default. I've
switched runrest_stop_message and runrest_ignore_message to this, since
this is what you need to do to override defaults (in fact there was a
non-working snippet here following 80dc71225541). I've also tried to
clarify that ^= matters for the interaction of the two.