feat: make AK start standing on exit
It's basically a spoiler that an exit always generates near the AK
starting spawn.  If the player doesn't know that, they'll waste a fair
bit of piety using Depart the Abyss.  That could be fine, except that
now that means their starting escape mechanism, Bend Space, will be
offline for the first few levels of the Dungeon, defeating the purpose
of bumping AK piety to make it available.

We could make the exit just generate in LOS, but there's nothing
really tactically interesting about walking a few squares and maybe
getting very unlucky, so just have AK start out standing on the exit.