feat: Simplify chances for special rooms
For the `special_room` vault, currently there's a 0.5% chance to place
this vault in Slime, a 1% chance to place this in Orc, Spider, Crypt,
and Pan, and a 3% chance to place it elsewhere. There's no reason to
have different yet extremely similar chances for this vault among the
non-extended branches. This commit uses a 2% chance for non-extended
branches and 1% for the Hells and Pan. This will reduce the frequency of
special rooms somewhat, but the vault does have `allow_dup` and hence
still has a decent chance of being seen one or more times for each
complete game.

The 1% chance for extended branches isn't quite justified either.
Arguably it might have some impact on balance for the Hells. For Pan we
have potentially infinite levels, so it may be good to reduce its
frequency. I'll keep it at 1% in these branches and leave any adjustment
to people wanting to tweak extended.