Make allies default to attacking
One of the most confusing and annoying things for players about allies
is that, by default, they will not attack enemies. If you or your ally
is hit by an enemy, or if you attack an enemy or order an attack, or if
you tell your allies to wait/guard, they'll attack; otherwise they'll
follow you uselessly. This feels buggy to players - understandably!

Change allies to default to attacking enemies instead. This is a slight
buff to summoning allies out of sight of enemies & a nerf to trying to
sneak around while hanging out with allies, but hopefully is a net
improvement to the player experience.