fix: tile sheet tweaks for better packing
This is basically an emergency patch to get the player.png tile sheet
size below height 2048, which won't work as a long term solution for the
relevant issues but can be back-ported to 0.28 at least. If the tile
sheet size exceeds 2048, on local tiles, it requires a 1024x4096 size
texture atlas. Any modern hardware will support this, but a non-trivial
amount of older or cheaper hardware won't[1]. Before 0.28, we had been
able to keep most tile sheets to 1024x1024 with the exception of
wall.png and player.png, both at 1024x2048. This is still not ideal
(https://crawl.develz.org/mantis/view.php?id=11378), but this is
something like 1% of graphics cards these days, as opposed to 6% for not
supported 1024x4096. Various additions in 0.28 pushed player.png up
across this threshold.

In the long run, it would be much better if tile sheets were
automatically split up and the implementation weren't tied to the size
of the tile sheet. However, the tech cost for doing this is really
painful, as the local tiles buffer implementation is really deeply tied
to the assumption that each texture atlas is stored in a single 2D
texture. So since I was able to tweak things to get player.png down a
bit, I'm going with that for now.

[1] https://feedback.wildfiregames.com/report/opengl/feature/GL_MAX_TEXTURE_SIZE

(cherry picked from commit b366d0e2d19ed67b3b9367b681a56fbbc7822b46)