New item: scroll of poison
Inspired by the late and largely unlamented 'sea of fire' Ignis ability,
the scroll of poison fills every unoccupied space on the screen with
poison clouds for 10-20 turns. Early on, it can be a way to escape enemies
or to attack them from range with impunity, especially with a source of
rPois or some !curing. Later, it can damage enemies reckless enough to
advance through,  block more dangerous clouds, reveal invisible enemies,
be ignited with Ignite Poison, trick enemies into walking into other clouds
you've placed... and so on. It also can amuse Xom.

It's closest to the condenser vane in use, but unlike the vane, it doesn't
place clouds directly on enemies, place non-poison clouds, scale with Evo, or
recharge.