fix: improve arena cancel/end behavior
* Remove all delays on cancelling, these are just confusing.
* Show popup if any trials have occurred, even if no explicit t:1 flag
  has been set. I guess it may not be super informative with only one
  trial, but there's no other good way to see the log if you want it.
  This includes even the cancellation case, but only if at least one
  complete trial has occurred.
* Add some more messaging around cancelling.

I wonder if there are interesting stats the final popup could show?

While testing this I noticed all sorts of other flaky behavior.
Sampling: Sometimes cancelling still results in a delay with a blank
screen. In console when I hit space during the first trial, there is a
redraw that wipes out half the banner. Sometimes the cursor shows up in
an ugly way (but sometimes it doesn't).