feat: Some adjustments to unrand vaults
dilly_guarded_unrand_throatcut: Scale the monsters better with depth.
Adjust the skeleton placement so that it's spread within each of the two
rooms, so that both rooms always get skeletons.

dilly_guarded_unrand_morg: Adjust monster set and placement to better
scale with depth. Use spectral things and simulacra and a wight band for
the single "boss" monster on D:5-7, and use either skeletal warrior or
less commonly a death knight for the boss monster on D:8-D:9. Give the
dead trees their usual custom description, and give them special
coloring for console.

dilly_guarded_unrand_devastator: Rework the monster list to better
connect with the themes of Devastator, and improve scaling with depth.
Use ogres as the basic threat, using normal ones for D:4-6 and
two-headed ogres for D:7-12. Introduce one gargoyles for D:7-D:9, adding
more for D:10-12, also adding a boulder beetle for that same range.
Place the unrand centrally in the vault, and make all statues use the
"orb" tile, since that visually best matches the theme.

Closes #2527.