fix: don't parse bad channels as no channel in runrest rc lines
This was happening because the crawl code returns nil on the error case
for parsing a channel name, which is then passed to crawl.message_filter
as a channel arg. Since this function supports channel as an optional
arg, nil maps to no channel (-1). So the gammafunk rc line `ignore
+= pray:` actually sets a runrest_ignore for basically any message.
After this change, the code errors and prevents the line from being used
at all, in line with what happens in channel parsing in initfile.cc.

Also fixes a buggy ignore involving Jiyva. (Though probably this should
only look for one word channel names?)

Note that this doesn't fix a larger issue leading to this commit, which
is that it is still possible to write runrest_ignore_message lines that
trip the buggy rest ASSERT under the right circumstances.