fix: hardcode an activity interrupt for barb removal
No rc options should be able to trigger an infinite rest, but currently
because of the way resting is implemented, if the regular interrupt
conditions are satisfied before the barb duration is cleared, only a
message interrupt (easily disableable) remains to stop the rest, making
it easy to force an infinite rest. In the past this has typically
happened by accident (e.g. as in 8ed1089bbe3), but it could also be used
to force a crash on demand and reset levels.

This commit takes a slightly hacky approach, and sends a full hp
interrupt explicitly when barbs are removed. It doesn't seem worth
adding a specific barb interrupt. But as long as barbs' interaction with
resting is hardcoded (see player::is_sufficiently_rested), something
hardcoded needs to happen on the end of a barb duration.