fix: ignore some disabled interrupts in rest checks
Before this commit, disabling rest interrupts could result in infinite
rests (and then a crash on demand in code attempting to detect them).
This commit changes things so that if hp or mp interrupts are disabled,
they are not factored into whether a rest can be started -- so if both
are disabled, resting actually becomes impossible (and a wait will be
started instead). That is, a rest that can't be interrupted is never
started. I think this may still have a weird interaction with
ancestor_hp, which I don't quite understand and haven't changed.